refactor(contracts): extract contract fetching logic into custom hook
- Created useContracts hook to handle contract data fetching and state management
- Refactored ContractsPage to use the new hook while maintaining existing functionality
- Improved code organization by separating data fetching concerns
- Reduced component complexity without changing behavior

This change makes the contract fetching logic reusable and the ContractsPage component more focused on presentation.
